+++ src/pages/Boards/index.jsx
@@ -1,0 +1,1 @@
+// Board List
