+++ src/pages/User/_id.jsx
@@ -1,0 +1,1 @@
+// User Details
